genomic/requirements.txt
django-filter
Django>=4.2
djangorestframework>=3.14
drf-spectacular>=0.27
django-cors-headers>=4.3
mysqlclient>=2.2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31